+++ botx/__init__.py
@@ -19,7 +19,6 @@ from botx.client.chats_api.exceptions import (
     ChatCreationProhibitedError,
     InvalidUsersListError,
 )
-from botx.client.chats_api.list_chats import ChatListItem
 from botx.client.exceptions.callbacks import (
     BotXMethodFailedCallbackReceivedError,
     CallbackNotReceivedError,
@@ -40,11 +39,10 @@ from botx.client.notifications_api.exceptions import (
     FinalRecipientsListEmptyError,
     StealthModeDisabledError,
 )
-from botx.client.notifications_api.markup import BubbleMarkup, Button, KeyboardMarkup
 from botx.client.users_api.exceptions import UserNotFoundError
 from botx.models.attachments import OutgoingAttachment
 from botx.models.bot_account import BotAccount
-from botx.models.chat import Chat
+from botx.models.chats import Chat, ChatListItem
 from botx.models.enums import ChatTypes, ClientPlatforms, MentionTypes, UserKinds
 from botx.models.message.entities import Mention
 from botx.models.message.incoming_message import (
@@ -53,6 +51,7 @@ from botx.models.message.incoming_message import (
     UserDevice,
     UserEventSender,
 )
+from botx.models.message.markup import BubbleMarkup, Button, KeyboardMarkup
 from botx.models.method_callbacks import BotAPIMethodFailedCallback
 from botx.models.status import BotMenu, StatusRecipient
 from botx.models.system_events.added_to_chat import AddedToChatEvent
@@ -61,6 +60,7 @@ from botx.models.system_events.deleted_from_chat import DeletedFromChatEvent
 
 __all__ = (
     "BotMenu",
+    "ChatListItem",
     "MentionTypes",
     "OutgoingAttachment",
     "IncomingMessage",
@@ -89,7 +89,6 @@ __all__ = (
     "Chat",
     "ChatCreationError",
     "ChatCreationProhibitedError",
-    "ChatListItem",
     "ChatNotFoundError",
     "FileDeletedError",
     "FileMetadataNotFound",
