+++ botx/client/users_api/user_from_search.py
@@ -1,40 +1,15 @@
-from dataclasses import dataclass
 from typing import List, Optional
 from uuid import UUID
 
 from pydantic import Field
 
-from botx.shared_models.api_base import VerifiedPayloadBaseModel
+from botx.models.api_base import VerifiedPayloadBaseModel
+from botx.models.users import UserFromSearch
 
 try:
     from typing import Literal
 except ImportError:
     from typing_extensions import Literal  # type: ignore  # noqa: WPS440
-
-
-@dataclass
-class UserFromSearch:
-    """User from search.
-
-    Attributes:
-        huid: User huid.
-        ad_login: User AD login.
-        ad_domain: User AD domain.
-        username: User name.
-        company: User company.
-        company_position: User company position.
-        department: User department.
-        emails: User emails.
-    """
-
-    huid: UUID
-    ad_login: Optional[str]
-    ad_domain: Optional[str]
-    username: str
-    company: Optional[str]
-    company_position: Optional[str]
-    department: Optional[str]
-    emails: List[str]
 
 
 class BotXAPISearchUserResult(VerifiedPayloadBaseModel):
