+++ botx/models/message/entities.py
@@ -4,13 +4,13 @@ from uuid import UUID
 
 from pydantic import Field, validator
 
+from botx.models.api_base import VerifiedPayloadBaseModel
 from botx.models.enums import (
     BotAPIEntityTypes,
     BotAPIMentionTypes,
     MentionTypes,
     convert_mention_type_to_domain,
 )
-from botx.shared_models.api_base import VerifiedPayloadBaseModel
 
 try:
     from typing import Literal
