+++ botx/models/status.py
@@ -2,9 +2,9 @@ from dataclasses import asdict, dataclass
 from typing import Any, Dict, List, NewType, Optional
 from uuid import UUID
 
+from botx.models.api_base import VerifiedPayloadBaseModel
 from botx.models.enums import APIChatTypes, ChatTypes, convert_chat_type_to_domain
 from botx.models.message.incoming_message import IncomingMessage
-from botx.shared_models.api_base import VerifiedPayloadBaseModel
 
 try:
     from typing import Literal
