+++ botx/models/system_events/added_to_chat.py
@@ -4,14 +4,14 @@ from uuid import UUID
 
 from pydantic import Field
 
+from botx.models.api_base import VerifiedPayloadBaseModel
 from botx.models.base_command import (
     BotAPIBaseCommand,
     BotAPIChatContext,
     BotCommandBase,
 )
-from botx.models.chat import Chat
+from botx.models.chats import Chat
 from botx.models.enums import BotAPICommandTypes, convert_chat_type_to_domain
-from botx.shared_models.api_base import VerifiedPayloadBaseModel
 
 try:
     from typing import Literal
